VBox/pci.h: int32_t -> size_t, fixes compile warning (and a size being a signed integer doesn't make much sense), bugref:10528
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@101466 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ include/VBox/pci.h
@@ -804,7 +804,7 @@ struct PCIBusAddress
     }
 
     /** Create string representation of this PCI address. */
-    bool format(char* szBuf, int32_t cBufSize)
+    bool format(char* szBuf, size_t cBufSize)
     {
         if (cBufSize < (/* bus */ 2 + /* : */ 1 + /* device */ 2 + /* . */ 1 + /* function*/ 1 + /* \0 */1))
             return false;
